requirements.txt
apache-airflow==2.8.0
mlflow==2.9.2
dvc==3.37.0
pandas==2.1.4
scikit-learn==1.3.2
numpy==1.26.3
Pillow==10.1.0
matplotlib==3.8.2
# PyTorch CPU-only (plus leger)
torch==2.1.2
torchvision==0.16.2

